Change to be using React Functional components
