+++ src/components/Footer.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Container, Icon } from 'semantic-ui-react'
 
-const Footer = () => (
+const Footer: React.FC = () => (
     <Container textAlign="center">
         Footer
         <Icon name="react" />
