+++ src/components/Home.tsx
@@ -1,10 +1,10 @@
-import React from 'react';
-import { Label } from 'semantic-ui-react';
+import React from 'react'
+import { Label } from 'semantic-ui-react'
 
-const Home = () => {
+const Home: React.FC = () => {
   return (
     <Label>Hello World of React and Webpack!</Label>
   );
 };
 
-export default Home;
+export default Home
